refactor(serverless): split pub/sub & s3 logic into separate crates
