+++ .changeset/nice-pans-train.md
@@ -1,0 +1,5 @@
+---
+'@lagon/serverless': patch
+---
+
+Remove useless LAGON_WORKERS env variable
