+++ .changeset/twelve-sheep-type.md
@@ -1,0 +1,5 @@
+---
+'@lagon/serverless': patch
+---
+
+Split pub/sub & s3 downloading logic into separate crates
