+++ crates/serverless/.env.example
@@ -2,7 +2,6 @@ LAGON_ROOT_DOMAIN=lagon.dev
 LAGON_REGION=local
 LAGON_ISOLATES_CACHE_SECONDS=60
 LAGON_LISTEN_ADDR=0.0.0.0:4000
-LAGON_WORKERS=4
 
 DATABASE_URL=mysql://root:mysql@localhost:3306/lagon
 REDIS_URL=redis://localhost:6379
