+++ crates/serverless/Cargo.toml
@@ -10,24 +10,21 @@ tokio-util = { version = "0.7.7", features = ["rt"] }
 lagon-runtime = { path = "../runtime" }
 lagon-runtime-http = { path = "../runtime_http" }
 lagon-runtime-isolate = { path = "../runtime_isolate" }
-lagon-serverless-logger = { path = "../serverless_logger" }
 lagon-runtime-utils = { path = "../runtime_utils" }
+lagon-serverless-logger = { path = "../serverless_logger" }
+lagon-serverless-downloader = { path = "../serverless_downloader" }
+lagon-serverless-pubsub = { path = "../serverless_pubsub" }
 flume = "0.10.14"
 mysql = "23.0.1"
 dotenv = "0.15.0"
-rust-s3 = "0.32"
-redis = { version = "0.22.3", features = ["tokio-native-tls-comp", "tokio-comp"] }
 serde_json = "1.0"
 metrics = "0.20.1"
 metrics-exporter-prometheus = { version = "0.11.0", default-features = false, features = ["http-listener"] }
 log = { version = "0.4.17", features = ["std", "kv_unstable", "kv_unstable_serde"] }
 lazy_static = "1.4.0"
-rand = { version = "0.8.5", features = ["std_rng"] }
 anyhow = "1.0.69"
 tokio-cron-scheduler = "0.9.4"
-uuid = "1.3.0"
 dashmap = "5.4.0"
-async-trait = "0.1.66"
 futures = "0.3.26"
 
 [build-dependencies]
