+++ crates/serverless/tests/deployments.rs
@@ -1,10 +1,9 @@
 use anyhow::Result;
 use dashmap::DashMap;
 use lagon_runtime_utils::Deployment;
-use lagon_serverless::{
-    deployments::{downloader::FakeDownloader, pubsub::FakePubSub},
-    serverless::start,
-};
+use lagon_serverless::serverless::start;
+use lagon_serverless_downloader::FakeDownloader;
+use lagon_serverless_pubsub::FakePubSub;
 use serial_test::serial;
 use std::{
     collections::{HashMap, HashSet},
