+++ crates/serverless/tests/errors.rs
@@ -4,10 +4,9 @@ use lagon_runtime_utils::{
     response::{PAGE_403, PAGE_404, PAGE_500, PAGE_502},
     Deployment,
 };
-use lagon_serverless::{
-    deployments::{downloader::FakeDownloader, pubsub::FakePubSub},
-    serverless::start,
-};
+use lagon_serverless::serverless::start;
+use lagon_serverless_downloader::FakeDownloader;
+use lagon_serverless_pubsub::FakePubSub;
 use serial_test::serial;
 use std::{
     collections::{HashMap, HashSet},
