+++ crates/serverless/tests/pubsub.rs
@@ -1,13 +1,9 @@
 use anyhow::Result;
 use dashmap::DashMap;
 use lagon_runtime_utils::response::{PAGE_403, PAGE_404};
-use lagon_serverless::{
-    deployments::{
-        downloader::FakeDownloader,
-        pubsub::{FakePubSub, PubSubMessage, PubSubMessageKind},
-    },
-    serverless::start,
-};
+use lagon_serverless::serverless::start;
+use lagon_serverless_downloader::FakeDownloader;
+use lagon_serverless_pubsub::{FakePubSub, PubSubMessage, PubSubMessageKind};
 use serial_test::serial;
 use std::{sync::Arc, time::Duration};
