+++ crates/serverless_pubsub/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "lagon-serverless-pubsub"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.69"
+async-trait = "0.1.66"
+futures = "0.3.26"
+redis = { version = "0.22.3", features = ["tokio-native-tls-comp", "tokio-comp"] }
+log = { version = "0.4.17", features = ["std", "kv_unstable", "kv_unstable_serde"] }
+flume = "0.10.14"
